+++ AndroidManifest.xml
@@ -81,8 +81,7 @@
 			/> 
 			      	
 			      	
-      	<activity android:name="org.torproject.android.ui.wizard.PromoAppsActivity" android:exported="false"/>
-      	<activity android:name="org.torproject.android.ui.wizard.ChooseLocaleWizardActivity" android:exported="false"/>
+      	<activity android:name="org.torproject.android.ui.PromoAppsActivity" android:exported="false"/>
       	
       	
       	<activity android:name=".settings.SettingsPreferences"  android:label="@string/app_name"/>
