+++ res/layout/layout_promo_apps.xml
