--- res/layout/layout_wizard.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="fill_parent"
-    android:layout_height="wrap_content">
-	<ScrollView android:id="@+id/helpscrollview" 
-		android:orientation="vertical"
-		android:layout_width="fill_parent" 
-		android:layout_height="fill_parent">    
-		<LinearLayout
-		    android:orientation="vertical"
-		    android:layout_width="fill_parent"
-		    android:layout_height="fill_parent">
-			<TextView android:text=""
-				android:layout_width="fill_parent" 
-				android:layout_height="wrap_content"
-				android:textColor="#ffffff" />				
-			</LinearLayout>
-		</ScrollView>			
-</LinearLayout>
-
-
