--- res/layout/layout_wizard_locale.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical" >
-            
-<RelativeLayout 
-	android:layout_width="fill_parent" 
-	android:layout_height="wrap_content" 
-	android:id="@+id/relativeLayout1" 
-	android:layout_gravity="center_horizontal"
-	android:background="#000000">
-         <TextView android:textSize="10pt" 
-        	android:textColor="#FFFFFF"
-        	android:layout_alignParentTop="true"
-       	 	android:layout_height="wrap_content" 
-        	android:id="@+id/WizardLocaleTitle" 
-        	android:layout_width="wrap_content" 
-        	android:fadingEdge="vertical" 
-        	android:text="@string/wizard_locale_title" 
-        	android:singleLine="true" 
-          	android:padding="20px" android:layout_centerInParent="true"> 
-        </TextView>
-    </RelativeLayout>    
-<TableLayout android:id="@+id/TableLayout01" android:stretchColumns="*" android:layout_height="fill_parent" android:layout_width="wrap_content">
-
-<TableRow android:layout_margin="10dip" android:id="@+id/TableRow01" android:layout_weight="1" android:layout_width="fill_parent" android:layout_height="fill_parent">
-  
-
-<LinearLayout
-			android:background="#575757"
-		    android:orientation="vertical"
-		    android:layout_width="fill_parent"
-		    android:layout_height="fill_parent"
-		    android:layout_weight="1" 
-		    android:padding="3px">
-<TextView android:text="@string/wizard_locale_msg" android:layout_width="wrap_content" android:textColor="#ffffff" android:layout_height="wrap_content"  android:textSize="8pt" android:id="@+id/WizardTextBody1" android:padding="20dip"></TextView>
-<ListView 
-    android:id="@+id/wizard_locale_list"
-    android:layout_width="fill_parent"
-	android:layout_height="wrap_content"
-	android:layout_weight="1"
-    >    
-</ListView>
-
-
-			</LinearLayout>
-		
-</TableRow>
-<TableRow android:background="#000000" android:layout_marginTop="10dip" android:paddingTop="10dip" android:id="@+id/TableRow01" android:textColor="#00ff00" android:layout_width="fill_parent" android:layout_height="30px">
-    	<Button android:text="@string/btn_next" android:id="@+id/btnWizard2"  android:layout_weight="1" android:layout_width="fill_parent" android:layout_height="wrap_content"></Button>	    
-</TableRow>			
-</TableLayout>
-</LinearLayout>
-
-
-
